requirements.txt
requests
bs4
lxml
flask
pycryptodome
